Make the logging less verbose
