+++ .gitignore
@@ -1,3 +1,4 @@
 # [tag:gitignore] Keep this in sync with [ref:excluded_input_paths].
+.data
 /artifacts/
 /target/
